separate app from main
